Adjust buttons
The buttons weren't resizing properly so I had to change the CSS in the
button for section one. I had set a hard limit on the height and width
so when I tried adjusting the font size the words would overflow. I got
rid of the height and width and used padding instead.